third_party/blink/web_tests/fast/scrolling/scroll-into-view-collapsed-div-expected.txt

Focus Textbox Collapse Div With Textbox
Focus Textbox Collapse Div With Textbox

Tests that scrollIntoViewIfNeeded works correctly if the enclosing div has no height.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS window.scrollY is 190
PASS window.scrollY is 190
PASS window.scrollX is 0
PASS window.scrollX is 0
